% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.BASGRABGC.R
\name{met2model.MODEL}
\alias{met2model.MODEL}
\alias{met2model.BASGRABGC}
\title{Write MODEL met files}
\usage{
met2model.BASGRABGC(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE
)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{overwrite}{logical: replace output files if they already exist?}
}
\value{
OK if everything was succesful.
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Rob Kooper
}
