% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.BASGRABGC.R
\name{write.config.MODEL}
\alias{write.config.MODEL}
\alias{write.config.BASGRABGC}
\title{Write MODEL configuration files}
\usage{
write.config.BASGRABGC(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for MODEL for given run
}
\description{
Writes a MODEL config file.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Rob Kooper
}
